openclip-docker-inference/requirements.txt
boto3
sagemaker-inference
open_clip_torch
botocore
